+++ ff/src/fields/models/fp/mod.rs
@@ -182,7 +182,7 @@ pub type Fp832<P> = Fp<P, 13>;
 
 impl<P, const N: usize> Fp<P, N> {
     /// Construct a new prime element directly from its underlying
-    /// [`BigInt`] data type.
+    /// [`struct@BigInt`] data type.
     #[inline]
     pub const fn new(element: BigInt<N>) -> Self {
         Self(element, PhantomData)
@@ -348,7 +348,6 @@ impl<P: FpConfig<N>, const N: usize> Field for Fp<P, N> {
 impl<P: FpConfig<N>, const N: usize> PrimeField for Fp<P, N> {
     type BigInt = BigInt<N>;
     const MODULUS: Self::BigInt = P::MODULUS;
-    const GENERATOR: Self = P::GENERATOR;
     const MODULUS_MINUS_ONE_DIV_TWO: Self::BigInt = P::MODULUS.divide_by_2_round_down();
     const MODULUS_BIT_SIZE: u32 = P::MODULUS.const_num_bits();
     const TRACE: Self::BigInt = P::MODULUS.two_adic_coefficient();