+++ test-curves/src/bls12_381/tests.rs
@@ -6,8 +6,11 @@ use crate::bls12_381::{g1, Fq, Fq2, Fq6, FqConfig, Fr, FrConfig, G1Affine, G1Pro
 use ark_algebra_test_templates::{
     curves::*, fields::*, generate_field_test, generate_g1_test, groups::*, msm::*,
 };
-use ark_std::ops::{AddAssign, MulAssign, SubAssign};
-use ark_std::{rand::Rng, test_rng};
+use ark_std::{
+    ops::{AddAssign, MulAssign, SubAssign},
+    rand::Rng,
+    test_rng,
+};
 
 generate_field_test!(bls12_381; fq2; fq6; mont(6, 4); );
 generate_g1_test!(bls12_381; curve_tests; sw_tests;);